Удаление таблиц и новый обработчик

+++ src/main/java/ru/yandex/practicum/filmorate/repository/UserRepository.java
@@ -5,6 +5,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 
 import ru.yandex.practicum.filmorate.exception.EntityUpdateErrorException;
+import ru.yandex.practicum.filmorate.exception.NotFoundIssueException;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -35,7 +36,7 @@ public class UserRepository extends BaseRepository<User> {
 
     public User saveUser(User user) {
         String query = "insert into users (email, login, name, birthday)"
-                + "values (?, ?, ?, ?)";
+                       + "values (?, ?, ?, ?)";
         int id = insert(query,
                 user.getEmail(),
                 user.getLogin(),
@@ -57,13 +58,18 @@ public class UserRepository extends BaseRepository<User> {
                 user.getId());
 
         if (result == 0) {
-            throw new EntityUpdateErrorException("Не удалось обновить пользователя");
+            throw new NotFoundIssueException("Пользователь с id = " + user.getId() + " не найден");
         }
 
-        return getUserById(user.getId()).get();
+        return getUserById(user.getId()).orElseThrow();
     }
 
     public void deleteUser(int id) {
+        Optional<User> user = getUserById(id);
+        if (user.isEmpty()) {
+            throw new NotFoundIssueException("Пользователь с id = " + id + " не найден");
+        }
+
         String queryEvents = "delete from events where user_id = ?";
         delete(queryEvents, id);
 
@@ -72,6 +78,15 @@ public class UserRepository extends BaseRepository<User> {
 
         String queryFriends2 = "delete from user_friends where friend_id = ?";
         delete(queryFriends2, id);
+
+        String queryLikes = "delete from likes where user_id = ?";
+        delete(queryLikes, id);
+
+        String queryReviews = "delete from reviews where user_id = ?";
+        delete(queryReviews, id);
+
+        String queryReviewLikes = "delete from review_likes where user_id = ?";
+        delete(queryReviewLikes, id);
 
         String queryUserS = "delete from users where id = ?";
         delete(queryUserS, id);
